+++ FastFood.Application/Ingredient/Command/IngredientDto.cs
@@ -4,7 +4,7 @@
     {
         public string Name { get; set; } = default!;
         public string Description { get; set; } = default!;
-        public int Prize { get; set; }
+        public decimal Prize { get; set; }
         public bool IsRequired { get; set; }
     }
 }
